Get app group IDs from entitlements
A bunch of people have tried to build the app from source and got confused when it crashed, because they customized the app group but didn't update the preprocessor macro. Now the app reads its app group off the mach-o on disk.

This also fixes AltStore, which has to change app group names too.

Fixes #650