Fix poll() on a tty not always stopping on ctrl-c
The tty would be woken up on a signal-generating ^C even though there
was no new input. This would make the poll() wake up, but it got a byte
from the pipe instead of an EINTR, and would just start waiting again.

* Don't wake up the tty when there is no new input, such as when a
signal comes in. This ensures that kqueue will return EINTR in this case
instead of the byte from the pipe. It's also just more efficient to
avoid unneeded wakeups.

* Break from the poll loop if nothing is ready and there's a signal
waiting. This is a more resilient fix and should handle any other
similar spurious wakeup situations.

Fixes #803