Redo the GitHub actions for TestFlight
Now makes different builds for App Store and sideloading, where the App
Store build is 13gb and the sideloading build is not. Uploads all builds
and dSYMs to GitHub releases, except the App Store ipa because it's
13gb. Release is marked "prerelease", so it doesn't actually go out to
AltStore until it's edited to be not prerelease. TestFlight build is
also not distributed externally (though the changelog is updated).

Almost completely untested. Let's see how it goes!