Update hterm to 1.91
This pulls in a bug fix for iOS 14.5. hterm used an old nonstandard version of the
Intl.Segmenter API that was unimplemented on Safari and polyfilled
instead. In 14.5, Safari implemented the finished version of the
standard, and hterm broke because the API existed but not in the way it
expected.