Much improved pty close handling.  Fixes mosh-server, dtach.
Many pty-using programs use wait() or SIGCHLD to detect child
termination, but some expect the pty master to close and return EOF
when all children close the slave pty.  This did not work on iSH.

If the slave pty was closed by the last program exiting, the program on the
master would not receive any read/poll wakeups, and the master pty
would not be hung up.  Reference counting wasn't working properly when
the pty was in a process group or session.  The fix for that is a bit
simplistic, but it's a big improvement.

This also improves the tty/pty code separation and fixes some other
minor things.