Promote hostname override to a friendly preference
Users really hate that we put "localhost" in their prompt as a result of
iOS 16 anti-fingerprinting obfuscating the device name. We already have
a mechanism for overriding this that we use for App Store screenshots,
so we might as well make it official and expose it for those that care.

The logic for what the actual uname hostname is as follows:

* If the user sets the preference, use that.
* On iOS <16, if the preference is unset, use what real uname returns.
* On iOS 16+, use UIDevice.name. This at least changes "localhost" to
something like "iPad".